+++ CHANGELOG.md
@@ -48,6 +48,9 @@ v2.0.0 (201x-xx-xx)
   namespaces for cluster backends (`Predis\Connection\Cluster`) and replication
   backends (`Predis\Connection\Replication`).
 
+- The method `Predis\Connection\AggregateConnectionInterface::getConnection()`
+  has been renamed to `getConnectionByCommand()`.
+
 - The methods `switchToMaster()` and `switchToSlave()` have been promoted to be
   part of `Predis\Connection\Replication\ReplicationInterface` while the method
   `switchTo($connection)` has been removed from it.